Refactor Copilot CLI loading and session management (#1730)
* Move Copilot CLI session out into own file

* Remvoe a few methods

* Updates
